Use table.pack to pass and return nil values
In Lua, table.unpack will stop at the first nil value it finds. This means that
wrap and async will not pass a complete set of arguments to the wrapped function
if one of the arguments is nil.

Use table.pack instead of {...} to count the number of packed values and provide
a range to table.unpack to unpack all the values.

Do the same thing to handle results with nil values returned from a coroutine.
Since table.remove also stops updating remaining table values when it encounters
a nil and does not accept a range, leave the coroutine result and return an
offset range of unpacked values.

Add some new tests to validate the expected behavior.